Change structure of domain packages to separate earthquake from coordinates
